pyproject.toml
[tool.poetry]
name = "compaipair"
version = "0.1.0"
description = "A simple CLI for pair programming with Google Generative AI APIs"
authors = ["Andrés Arango Pérez <andresap@protonmail.com>"]
license = "MIT"
include = [
    "data/**"
]

[tool.poetry.dependencies]
python = "^3.10"
google-generativeai = "^0.2.2"
click = "^8.1.7"
pytest = "^7.4.2"
tinydb = "^4.8.0"

[tool.poetry.dev-dependencies]
black = "^23.10.1"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
compai = "compaipair.main:compai"
